% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{tr_bill_rates}
\alias{tr_bill_rates}
\title{Return the daily treasury bill rates}
\usage{
tr_bill_rates(date = NULL)
}
\description{
Return the daily treasury bill rates
}
\references{
\url{https://home.treasury.gov/treasury-daily-interest-rate-xml-feed}
}
