% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield-curve.R
\name{tr_hqm}
\alias{tr_hqm}
\alias{tr_hqm_pars}
\title{Download Treasury Yield Curve Data}
\usage{
tr_hqm(x = c("average", "end-of-month"), year = NULL)

tr_hqm_pars(x = c("average", "end-of-month"))
}
\arguments{
\item{x}{\code{character(1)}. Either \code{"average"} or \code{"end-of-month"}.
Default is \code{"average"}.}

\item{year}{\code{integer(1)}. Year to download. Default is \code{NULL}.
If \code{NULL}, then all available years are downloaded.}
}
\value{
A \code{data.frame()} with the following columns: \code{yearmonth}, \code{maturity},
\code{yield}.
}
\description{
Download Treasury Yield Curve Data
}
\concept{yield curve}
