% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treasury-package.R
\docType{package}
\name{treasury-package}
\alias{treasury}
\alias{treasury-package}
\title{treasury: Client for US Treasury XML Feed and Published Data}
\description{
Download daily interest rates from the US Treasury XML feed. Leveraging \url{https://home.treasury.gov/treasury-daily-interest-rate-xml-feed}, this package serves as a wrapper, facilitating the retrieval of daily treasury rates across various categories, including par yield curves, treasury bills, long-term rates, and real yield curves. In addition, it provides access to the monthly published yield curve datasets, including the High Quality Market (HQM) corporate bond yield curve and the Treasury nominal and real coupon-issue (TNC, TRC) and breakeven inflation (TBI) curves.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://m-muecke.github.io/treasury/}
  \item \url{https://github.com/m-muecke/treasury}
  \item Report bugs at \url{https://github.com/m-muecke/treasury/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Mücke \email{muecke.maximilian@gmail.com} (\href{https://orcid.org/0009-0000-9432-9795}{ORCID})

Authors:
\itemize{
  \item Maximilian Mücke \email{muecke.maximilian@gmail.com} (\href{https://orcid.org/0009-0000-9432-9795}{ORCID})
}

}
\keyword{internal}
