% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest-rate.R
\name{tr_bill_rates}
\alias{tr_bill_rates}
\title{Daily treasury bill rates}
\usage{
tr_bill_rates(date = NULL)
}
\arguments{
\item{date}{\code{character(1)} or \code{numeric(1)} date in format yyyy or yyyymm.
If \code{NULL}, all data is returned. Default \code{NULL}.}
}
\value{
A \code{data.frame()} containing the rates or \code{NULL} when no entries were found.
}
\description{
These rates are indicative closing market bid quotations on the most recently
auctioned Treasury Bills in the over-the-counter market as obtained by the Federal
Reserve Bank of New York at approximately 3:30 PM each business day.
}
\examples{
\donttest{
# get data for a single month
tr_bill_rates("202201")
# or for the entire year
tr_bill_rates(2022)
}
}
\references{
\url{https://home.treasury.gov/treasury-daily-interest-rate-xml-feed}
}
\seealso{
Other interest rate: 
\code{\link{tr_long_term_rate}()},
\code{\link{tr_real_long_term}()},
\code{\link{tr_real_yield_curve}()},
\code{\link{tr_yield_curve}()}
}
\concept{interest rate}
