% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield-curve.R
\name{tr_hqm}
\alias{tr_hqm}
\alias{tr_hqm_pars}
\title{The Treasury High Quality Market (HQM) Corporate Bond Yield Curve}
\source{
\url{https://home.treasury.gov/data/treasury-coupon-issues-and-corporate-bond-yield-curve/corporate-bond-yield-curve}
}
\usage{
tr_hqm(x = c("average", "end-of-month"), year = NULL)

tr_hqm_pars(x = c("average", "end-of-month"))
}
\arguments{
\item{x}{\code{character(1)}. Either \code{"average"} or \code{"end-of-month"}.
Default is \code{"average"}.}

\item{year}{\code{integer(1)}. Year to download. Default is \code{NULL}.
If \code{NULL}, then all available years are downloaded.}
}
\value{
A \code{data.frame()} with the following columns: \code{yearmonth}, \code{maturity},
\code{spot_rate}.
}
\description{
The Treasury High Quality Market (HQM) Corporate Bond Yield Curve
}
\concept{yield curve}
