% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{tr_long_term_rate}
\alias{tr_long_term_rate}
\title{Return the daily treasury long-term rates}
\usage{
tr_long_term_rate(date = NULL)
}
\description{
Return the daily treasury long-term rates
}
\references{
\url{https://home.treasury.gov/treasury-daily-interest-rate-xml-feed}
}
