% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{tr_real_long_term}
\alias{tr_real_long_term}
\title{Return the daily treasury real long-term rates}
\usage{
tr_real_long_term(date = NULL)
}
\description{
Return the daily treasury real long-term rates
}
\references{
\url{https://home.treasury.gov/treasury-daily-interest-rate-xml-feed}
}
