% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{tr_yield_curve}
\alias{tr_yield_curve}
\title{Return the daily treasury par yield curve rates}
\usage{
tr_yield_curve(date = NULL)
}
\description{
Return the daily treasury par yield curve rates
}
\references{
\url{https://home.treasury.gov/treasury-daily-interest-rate-xml-feed}
}
