% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest-rates.R
\name{tr_yield_curve}
\alias{tr_yield_curve}
\alias{tr_bill_rates}
\alias{tr_long_term_rate}
\alias{tr_real_yield_curve}
\alias{tr_real_long_term}
\title{Returns the daily treasury interest rate}
\usage{
tr_yield_curve(date = NULL)

tr_bill_rates(date = NULL)

tr_long_term_rate(date = NULL)

tr_real_yield_curve(date = NULL)

tr_real_long_term(date = NULL)
}
\arguments{
\item{date}{\code{character(1)} or \code{numeric(1)} date in format yyyy or yyyymm.
If \code{NULL}, all data is returned. Default \code{NULL}.}
}
\value{
A \code{data.frame()} containing the rates or \code{NULL} when no entries were found.
}
\description{
Returns the daily treasury interest rate
}
\examples{
\donttest{
# get data for a single month
tr_yield_curve("202201")
# or for the entire year
tr_yield_curve(2022)
}
}
\references{
\url{https://home.treasury.gov/treasury-daily-interest-rate-xml-feed}
}
